# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ChangePasswordRequest:
    properties:
      confirm_password:
        type: string
      current_password:
        type: string
      new_password:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - confirm_password
    - current_password
    - new_password
    type: object
...
  dto.ErrorInfo:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  dto.Meta:
    properties:
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/dto.ErrorInfo'
      meta:
        $ref: '#/definitions/dto.Meta'
      success:
        type: boolean
    type: object
...
  dto.UpdateProfileRequest:
    properties:
      username:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - username
    type: object
  dto.WithdrawRequest:
    properties:
      password:
        type: string
      reason:
        type: string
    required:
    - password
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 게시판 API 서버입니다.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go Board API
  version: "1.0"
paths:
...
  /me:
    delete:
      consumes:
      - application/json
      description: 회원 탈퇴를 진행합니다
      parameters:
      - description: 탈퇴 확인 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.WithdrawRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 비밀번호 불일치
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 회원 탈퇴
      tags:
      - users
  /me/comments:
    get:
      description: 로그인한 사용자가 작성한 댓글 목록을 조회합니다
      parameters:
      - default: 1
        description: 페이지 번호
        in: query
        name: page
        type: integer
      - default: 10
        description: 페이지 크기
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: 인증 필요
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 내 댓글 목록
      tags:
      - users
  /me/password:
    put:
      consumes:
      - application/json
      description: 비밀번호를 변경합니다
      parameters:
      - description: 비밀번호 변경 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 비밀번호 불일치
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 비밀번호 변경
      tags:
      - users
  /me/posts:
    get:
      description: 로그인한 사용자가 작성한 게시글 목록을 조회합니다
      parameters:
      - default: 1
        description: 페이지 번호
        in: query
        name: page
        type: integer
      - default: 10
        description: 페이지 크기
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: 인증 필요
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 내 게시글 목록
      tags:
      - users
  /me/profile:
    get:
      description: 로그인한 사용자의 프로필을 조회합니다
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: 인증 필요
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 내 프로필 조회
      tags:
      - users
    put:
      consumes:
      - application/json
      description: 사용자 프로필을 수정합니다
      parameters:
      - description: 프로필 수정 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: 유효성 검사 실패
          schema:
            $ref: '#/definitions/dto.Response'
        "409":
          description: 사용자명 중복
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 프로필 수정
      tags:
      - users
...
  /users/search:
    get:
      description: 멘션 자동완성을 위한 사용자 검색입니다
      parameters:
      - description: 검색어 (2자 이상)
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - Bearer: []
      summary: 사용자 검색
      tags:
      - users
securityDefinitions:
  Bearer:
    description: 'JWT 인증 토큰. 형식: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...